requirements.txt
python-telegram-bot==21.9
apscheduler==3.10.4
python-dotenv==1.0.0
aiosqlite==0.20.0